Rename get-ingredients-names
